change proto type to standard 3

+++ Cheetah_Business/Data/CartableDTO.cs
@@ -7,10 +7,10 @@ public class CartableDTO
 {
     private CPersianClass? cPersianClass = new CPersianClass();
 
-    public Int32? PageNumber { get; set; }
+    public Int64? PageNumber { get; set; }
     public Int64? ERPCode { get; set; }
-    public Int32? PageSize { get; set; }
-    public Int32? TotalItems { get; set; }
+    public Int64? PageSize { get; set; }
+    public Int64? TotalItems { get; set; }
     public SimpleClassDTO? Tag { get; set; }
     public SimpleClassDTO? User { get; set; }
     public SimpleClassDTO? Requestor { get; set; }
